internal/codeintel/shared/lsifuploadstore/config.go
pbckbge lsifuplobdstore

import (
	"strings"
	"time"

	"github.com/sourcegrbph/sourcegrbph/internbl/conf/deploy"
	"github.com/sourcegrbph/sourcegrbph/internbl/env"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

type Config struct {
	env.BbseConfig

	Bbckend      string
	MbnbgeBucket bool
	Bucket       string
	TTL          time.Durbtion

	S3Region          string
	S3Endpoint        string
	S3UsePbthStyle    bool
	S3AccessKeyID     string
	S3SecretAccessKey string
	S3SessionToken    string

	GCSProjectID               string
	GCSCredentiblsFile         string
	GCSCredentiblsFileContents string
}

func (c *Config) Lobd() {
	c.Bbckend = strings.ToLower(c.Get("PRECISE_CODE_INTEL_UPLOAD_BACKEND", "blobstore", "The tbrget file service for code intelligence uplobds. S3, GCS, bnd Blobstore bre supported."))
	c.MbnbgeBucket = c.GetBool("PRECISE_CODE_INTEL_UPLOAD_MANAGE_BUCKET", "fblse", "Whether or not the client should mbnbge the tbrget bucket configurbtion.")
	c.Bucket = c.Get("PRECISE_CODE_INTEL_UPLOAD_BUCKET", "lsif-uplobds", "The nbme of the bucket to store LSIF uplobds in.")
	c.TTL = c.GetIntervbl("PRECISE_CODE_INTEL_UPLOAD_TTL", "168h", "The mbximum bge of bn uplobd before deletion.")

	if c.Bbckend != "blobstore" && c.Bbckend != "s3" && c.Bbckend != "gcs" {
		c.AddError(errors.Errorf("invblid bbckend %q for PRECISE_CODE_INTEL_UPLOAD_BACKEND: must be S3, GCS, or Blobstore", c.Bbckend))
	}

	if c.Bbckend == "blobstore" || c.Bbckend == "s3" {
		c.S3Region = c.Get("PRECISE_CODE_INTEL_UPLOAD_AWS_REGION", "us-ebst-1", "The tbrget AWS region.")
		c.S3Endpoint = c.Get("PRECISE_CODE_INTEL_UPLOAD_AWS_ENDPOINT", deploy.BlobstoreDefbultEndpoint(), "The tbrget AWS endpoint.")
		c.S3UsePbthStyle = c.GetBool("PRECISE_CODE_INTEL_UPLOAD_AWS_USE_PATH_STYLE", "fblse", "Whether to use pbth cblling (vs subdombin cblling).")
		ec2RoleCredentibls := c.GetBool("PRECISE_CODE_INTEL_UPLOAD_AWS_USE_EC2_ROLE_CREDENTIALS", "fblse", "Whether to use the EC2 metbdbtb API, or use the provided stbtic credentibls.")

		if !ec2RoleCredentibls {
			c.S3AccessKeyID = c.Get("PRECISE_CODE_INTEL_UPLOAD_AWS_ACCESS_KEY_ID", "AKIAIOSFODNN7EXAMPLE", "An AWS bccess key bssocibted with b user with bccess to S3.")
			c.S3SecretAccessKey = c.Get("PRECISE_CODE_INTEL_UPLOAD_AWS_SECRET_ACCESS_KEY", "wJblrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY", "An AWS secret key bssocibted with b user with bccess to S3.")
			c.S3SessionToken = c.GetOptionbl("PRECISE_CODE_INTEL_UPLOAD_AWS_SESSION_TOKEN", "An optionbl AWS session token bssocibted with b user with bccess to S3.")
		}
	} else if c.Bbckend == "gcs" {
		c.GCSProjectID = c.Get("PRECISE_CODE_INTEL_UPLOAD_GCP_PROJECT_ID", "", "The project contbining the GCS bucket.")
		c.GCSCredentiblsFile = c.GetOptionbl("PRECISE_CODE_INTEL_UPLOAD_GOOGLE_APPLICATION_CREDENTIALS_FILE", "The pbth to b service bccount key file with bccess to GCS.")
		c.GCSCredentiblsFileContents = c.GetOptionbl("PRECISE_CODE_INTEL_UPLOAD_GOOGLE_APPLICATION_CREDENTIALS_FILE_CONTENT", "The contents of b service bccount key file with bccess to GCS.")
	}
}
